switch to local library

+++ src/lib.rs
@@ -9,11 +9,11 @@ const BUFFER_SIZE: usize = 512;
 mod tests {
     extern crate winapi;
 
-    use self::winapi::ctypes::{c_void, c_ulong, c_long};
     use winapi::um::fileapi;
     use block_device::BlockDevice;
+    use core::ptr;
     use crate::volume::Volume;
-    use self::winapi::_core::ptr;
+    use self::winapi::ctypes::{c_void, c_ulong, c_long};
 
     const GENERIC_READ: c_ulong = 1 << 31;
     const FILE_SHARE_READ: c_ulong = 0x00000001;
@@ -80,14 +80,14 @@ mod tests {
     impl BlockDevice for Device {
         type Error = DeviceError;
 
-        fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error> {
+        fn read(&self, buf: &mut [u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error> {
             let mut len = 0;
             self.set_file_pointer(address as i32);
-            let res = self.read(buf, number_of_blocks, &mut len);
+            let res = self.read(buf, number_of_blocks as c_ulong, &mut len);
             if res { Ok(()) } else { Err(DeviceError::ReadError) }
         }
 
-        fn write(&self, buf: &[u8], address: usize, number_of_blocks: u32) -> Result<(), Self::Error> {
+        fn write(&self, buf: &[u8], address: usize, number_of_blocks: usize) -> Result<(), Self::Error> {
             unimplemented!()
         }
     }
